pyproject.toml
[tool.poetry]
name = "sample-celery"
version = "0.1.0"
description = ""
authors = ["espoir murhabazi <espoir.mur@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
celery = "4.4.7"
celery-redbeat = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
